Split mode class to a new file
