+++ src/main/scala/org/vim_jp/modal/MODalPlugin.scala
@@ -1,12 +1,9 @@
 package org.vim_jp.modal
 
-import org.bukkit.Bukkit
 import org.bukkit.Material
 import org.bukkit.NamespacedKey
-import org.bukkit.Server
 import org.bukkit.block.Block
 import org.bukkit.block.BlockFace
-import org.bukkit.block.data.Ageable
 import org.bukkit.command.Command
 import org.bukkit.command.CommandExecutor
 import org.bukkit.command.CommandSender
@@ -16,12 +13,11 @@ import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.block.BlockBreakEvent
 import org.bukkit.event.entity.ProjectileHitEvent
-import org.bukkit.inventory.ItemStack
-import org.bukkit.metadata.FixedMetadataValue
-import org.bukkit.metadata.MetadataValue
 import org.bukkit.persistence.PersistentDataType
 import org.bukkit.plugin.java.JavaPlugin
 import org.bukkit.scheduler.BukkitRunnable
+
+import org.vim_jp.modal.mode.FarmerMode
 
 object MODalPlugin:
   val MODE_DATA_KEY: String = "mode"
@@ -30,13 +26,13 @@ object MODalPlugin:
 class MODalPlugin extends JavaPlugin:
   outer =>
 
-  protected val modeDataKey = NamespacedKey(outer, MODalPlugin.MODE_DATA_KEY)
+  val modeDataKey = NamespacedKey(outer, MODalPlugin.MODE_DATA_KEY)
 
   override def onEnable(): Unit =
     val server = getServer
     val pluginManager = server.getPluginManager
     pluginManager.registerEvents(Kikori(), this)
-    pluginManager.registerEvents(Farmer(), this)
+    pluginManager.registerEvents(FarmerMode(this), this)
     pluginManager.registerEvents(ArrowWarp(), this)
     this.getCommand("change").setExecutor(CommandModal())
 
@@ -64,50 +60,6 @@ class MODalPlugin extends JavaPlugin:
       true
 
   // TODO: create TabCompleteEvent for the command
-
-  class Farmer extends Listener:
-    def seedOf(block: Block): Material =
-      block.getType.name match
-        case "WHEAT"    => Material.WHEAT_SEEDS
-        case "POTATOES" => Material.POTATO
-        case _          => null
-
-    @EventHandler
-    def onBlockBreak(event: BlockBreakEvent): Unit =
-      if event.isCancelled then return
-
-      val player = event.getPlayer
-
-      val container = player.getPersistentDataContainer()
-      val mode = container.get(outer.modeDataKey, PersistentDataType.STRING)
-      if mode != "farmer" then return
-
-      val block = event.getBlock: Block
-
-      val seed = seedOf(block)
-      if seed == null then return
-
-      val ageable = block.getBlockData match
-        case ageable: Ageable => ageable
-        case _                => return
-      if ageable.getAge != ageable.getMaximumAge then
-        event.setCancelled(true)
-        return
-
-      // consume seed
-      val item = ItemStack(seed)
-      val removed = player.getInventory().removeItem(item)
-      // (there's no seed)
-      if !removed.isEmpty then return
-
-      // set block to younuest state
-      val typ = block.getType
-      val loc = block.getLocation()
-      new BukkitRunnable {
-        override def run(): Unit =
-          val newBlock = loc.getBlock
-          newBlock.setType(typ)
-      }.runTask(outer)
 
   class Kikori extends Listener:
     def isLog(block: Block): Boolean =
